Text.Trifecta.Combinators

Description

Documentation

class (MonadPlus m, TokenParsing m) => DeltaParsing m where

Methods

line :: m ByteString

Retrieve the contents of the current line (from the beginning of the line)

position :: m Delta

slicedWith :: (a -> ByteString -> r) -> m a -> m r

run a parser, grabbing all of the text between its start and end points

rend :: DeltaParsing m => m Rendering

restOfLine :: DeltaParsing m => m ByteString

grab the remainder of the current line

sliced :: DeltaParsing m => m a -> m ByteString

run a parser, grabbing all of the text between its start and end points and discarding the original result

careting :: DeltaParsing m => m Caret

careted :: DeltaParsing m => m a -> m (Careted a)

spanning :: DeltaParsing m => m a -> m Span

spanned :: DeltaParsing m => m a -> m (Spanned a)

fixiting :: DeltaParsing m => m ByteString -> m Fixit

class (DeltaParsing m, HasDelta d) => MarkParsing d m | m -> d where

Methods

mark :: m d

mark the current location so it can be used in constructing a span, or for later seeking

release :: d -> m ()

Seek a previously marked location