Text.Trifecta.Delta

Description

A Delta keeps track of the cursor position of the parser, so it can be referred to later, for example in error messages.

Documentation

data Delta

Since there are multiple ways to be at a certain location, Delta captures all these alternatives as a single type.

Constructors

Columns !Int64 !Int64
( number of characters
, number of bytes )

Tab !Int64 !Int64 !Int64
( number of characters before the tab
, number of characters after the tab
, number of bytes )

Lines !Int64 !Int64 !Int64 !Int64
( number of newlines contained
, number of characters since the last newline
, number of bytes
, number of bytes since the last newline )

Directed !ByteString !Int64 !Int64 !Int64 !Int64
( current file name
, number of lines since the last line directive
, number of characters since the last newline
, number of bytes
, number of bytes since the last newline )

class HasDelta t where

Methods

delta :: t -> Delta

class HasBytes t where

Methods

bytes :: t -> Int64

prettyDelta :: Delta -> Doc AnsiStyle

Example: file.txt:12:34

nextTab :: Int64 -> Int64

Increment a column number to the next tabstop.

rewind :: Delta -> Delta

Rewind a Delta to the beginning of the line.

near :: (HasDelta s, HasDelta t) => s -> t -> Bool

Should we show two things with a Delta on the same line?

>>> near (Columns 0 0) (Columns 5 5)
True

>>> near (Lines 1 0 1 0) (Lines 2 4 4 2)
False

column :: HasDelta t => t -> Int64

Retrieve the character offset within the current line from this Delta.

columnByte :: Delta -> Int64

Retrieve the byte offset within the current line from this Delta.