Text.Trifecta.Diagnostic.Rendering.Prim

Description

Diagnostics rendering

The type for Lines will very likely change over time, so we can draw lit up control characters for ^Z, ^[, 0xff, etc. this will make for much nicer diagnostics when working with protocols!

Documentation

data Rendering

Constructors

Rendering

Fields

renderingDelta :: !Delta

renderingLineLen :: !Int

renderingLine :: Lines -> Lines

renderingOverlays :: Delta -> Lines -> Lines

nullRendering :: Rendering -> Bool

emptyRendering :: Rendering

class Source t where

Methods

source :: t -> (Int, Lines -> Lines)

rendering :: Source s => Delta -> s -> Rendering

create a drawing surface

class Renderable t where

Methods

render :: t -> Rendering

data Rendered a

Constructors

a :@ Rendering

Lower level drawing primitives

type Lines = Array (Int, Int) ([ScopedEffect], Char)

draw :: [ScopedEffect] -> Int -> Int -> String -> Lines -> Lines

ifNear :: Delta -> (Lines -> Lines) -> Delta -> Lines -> Lines

(.#) :: (Delta -> Lines -> Lines) -> Rendering -> Rendering