Text.Trifecta.Parser

Description

Documentation

newtype Parser a

Constructors

Parser

Fields

unparser :: forall r. (a -> Err -> It Rope r) -> (Err -> It Rope r) -> (a -> Set String -> Delta -> ByteString -> It Rope r) -> (TermDoc -> It Rope r) -> Delta -> ByteString -> It Rope r

manyAccum :: (a -> [a] -> [a]) -> Parser a -> Parser [a]

Parse Results

data Result a

Constructors

Success a

Failure TermDoc

Recoverable Parsing Errors

data Err

Constructors

Err

Fields

_reason :: Maybe TermDoc

_aux :: [TermDoc]

_expected :: Set String

class HasErr t where

Methods

err :: Lens t t Err Err

aux :: Lens t t [TermDoc] [TermDoc]

expected :: Lens t t (Set String) (Set String)

reason :: Lens t t (Maybe TermDoc) (Maybe TermDoc)

explain :: Rendering -> Err -> TermDoc

Feeding a parser more more input

data Step a

Constructors

StepDone !Rope a

StepFail !Rope TermDoc

StepCont !Rope (Result a) (Rope -> Step a)

feed :: Reducer t Rope => t -> Step r -> Step r

starve :: Step a -> Result a

stepParser :: Parser a -> Delta -> ByteString -> Step a

stepResult :: Rope -> Result a -> Step a

stepIt :: It Rope a -> Step a

Parsing

parseFromFile :: Show a => Parser a -> String -> IO (Maybe a)

parseFromFile p filePath runs a parser p on the input read from filePath using readFile. All diagnostic messages emitted over the course of the parse attempt are shown to the user on the console.

 main = do
   result <- parseFromFile numbers "digits.txt"
   case result of
     Nothing -> return ()
     Just a  -> print $ sum a

parseFromFileEx :: Show a => Parser a -> String -> IO (Result a)

parseFromFileEx p filePath runs a parser p on the input read from filePath using readFile. Returns all diagnostic messages emitted over the course of the parse and the answer if the parse was successful.

 main = do
   result <- parseFromFileEx (many number) "digits.txt"
   case result of
     Failure xs -> displayLn xs
     Success a  -> print (sum a)

parseByteString :: Show a => Parser a -> Delta -> ByteString -> Result a

parseByteString p delta i runs a parser p on i.

parseTest :: Show a => Parser a -> String -> IO ()