Text.Trifecta.Parser.Class

Description

Documentation

class (Alternative m, MonadPlus m) => MonadParser m where

Methods

try :: m a -> m a

Take a parser that may consume input, and on failure, go back to where we started and fail as if we didn't consume input.

labels :: m a -> Set String -> m a

liftIt :: It Rope a -> m a

Lift an operation from the primitive It monad

mark :: m Delta

mark the current location so it can be used in constructing a span, or for later seeking

unexpected :: MonadParser m => String -> m a

Used to emit an error on an unexpected token

line :: m ByteString

Retrieve the contents of the current line (from the beginning of the line)

skipMany :: m a -> m ()

A version of many that discards its input. Specialized because it can often be implemented more cheaply.

release :: Delta -> m ()

Seek back to previously marked location

satisfy :: (Char -> Bool) -> m Char

Parse a single character of the input, with UTF-8 decoding

satisfy8 :: (Word8 -> Bool) -> m Word8

Parse a single byte of the input, without UTF-8 decoding

highlightToken :: TokenHighlight -> m a -> m a

highlightToken is called internally in the token parsers. It delimits ranges of the input recognized by certain parsers that are useful for syntax highlighting. An interested monad could choose to listen to these events and construct an interval tree for later pretty printing purposes.

satisfyAscii :: MonadParser m => (Char -> Bool) -> m Char

restOfLine :: MonadParser m => m ByteString

grab the remainder of the current line

(<?>) :: MonadParser m => m a -> String -> m a

label a parser with a name

skipping :: MonadParser m => Delta -> m ()

slicedWith :: MonadParser m => (a -> ByteString -> r) -> m a -> m r

run a parser, grabbing all of the text between its start and end points

sliced :: MonadParser m => m a -> m ByteString

run a parser, grabbing all of the text between its start and end points and discarding the original result

rend :: MonadParser m => m Rendering