Text.Trifecta.Parser.Token

Documentation

module Text.Trifecta.Parser.Token.Class

module Text.Trifecta.Parser.Token.Combinators

module Text.Trifecta.Parser.Token.Identifier

decimal :: MonadParser m => m Integer

Parses a positive whole number in the decimal system. Returns the value of the number.

hexadecimal :: MonadParser m => m Integer

Parses a positive whole number in the hexadecimal system. The number should be prefixed with "0x" or "0X". Returns the value of the number.

octal :: MonadParser m => m Integer

Parses a positive whole number in the octal system. The number should be prefixed with "0o" or "0O". Returns the value of the number.