Text.Trifecta.Parser.Token.Combinators

Description

Documentation

charLiteral :: MonadTokenParser m => m Char

This lexeme parser parses a single literal character. Returns the literal character value. This parsers deals correctly with escape sequences. The literal character is parsed according to the grammar rules defined in the Haskell report (which matches most programming languages quite closely).

stringLiteral :: MonadTokenParser m => m String

This lexeme parser parses a literal string. Returns the literal string value. This parsers deals correctly with escape sequences and gaps. The literal string is parsed according to the grammar rules defined in the Haskell report (which matches most programming languages quite closely).

natural :: MonadTokenParser m => m Integer

This lexeme parser parses a natural number (a positive whole number). Returns the value of the number. The number can be specified in decimal, hexadecimal or octal. The number is parsed according to the grammar rules in the Haskell report.

integer :: MonadTokenParser m => m Integer

This lexeme parser parses an integer (a whole number). This parser is like natural except that it can be prefixed with sign (i.e. '-' or '+'). Returns the value of the number. The number can be specified in decimal, hexadecimal or octal. The number is parsed according to the grammar rules in the Haskell report.

double :: MonadTokenParser m => m Double

This lexeme parser parses a floating point value. Returns the value of the number. The number is parsed according to the grammar rules defined in the Haskell report.

naturalOrDouble :: MonadTokenParser m => m (Either Integer Double)

This lexeme parser parses either natural or a float. Returns the value of the number. This parsers deals with any overlap in the grammar rules for naturals and floats. The number is parsed according to the grammar rules defined in the Haskell report.

symbol :: MonadTokenParser m => ByteString -> m ByteString

Lexeme parser symbol s parses string s and skips trailing white space.

symbolic :: MonadTokenParser m => Char -> m Char

Lexeme parser symbolic s parses char s and skips trailing white space.

parens :: MonadTokenParser m => m a -> m a

Lexeme parser parens p parses p enclosed in parenthesis, returning the value of p.

braces :: MonadTokenParser m => m a -> m a

Lexeme parser braces p parses p enclosed in braces ('{' and '}'), returning the value of p.

angles :: MonadTokenParser m => m a -> m a

Lexeme parser angles p parses p enclosed in angle brackets ('<' and '>'), returning the value of p.

brackets :: MonadTokenParser m => m a -> m a

Lexeme parser brackets p parses p enclosed in brackets ('[' and ']'), returning the value of p.

semi :: MonadTokenParser m => m Char

Lexeme parser |semi| parses the character ';' and skips any trailing white space. Returns the string ";".

comma :: MonadTokenParser m => m Char

Lexeme parser comma parses the character ',' and skips any trailing white space. Returns the string ",".

colon :: MonadTokenParser m => m Char

Lexeme parser colon parses the character ':' and skips any trailing white space. Returns the string ":".

dot :: MonadTokenParser m => m Char

Lexeme parser dot parses the character '.' and skips any trailing white space. Returns the string ".".

semiSep :: MonadTokenParser m => m a -> m [a]

Lexeme parser semiSep p parses zero or more occurrences of p separated by semi. Returns a list of values returned by p.

semiSep1 :: MonadTokenParser m => m a -> m [a]

Lexeme parser semiSep1 p parses one or more occurrences of p separated by semi. Returns a list of values returned by p.

commaSep :: MonadTokenParser m => m a -> m [a]

Lexeme parser commaSep p parses zero or more occurrences of p separated by comma. Returns a list of values returned by p.

commaSep1 :: MonadTokenParser m => m a -> m [a]

Lexeme parser commaSep1 p parses one or more occurrences of p separated by comma. Returns a list of values returned by p.