Text.Trifecta.Rendering

Description

The type for Lines will very likely change over time, to enable drawing lit up multi-character versions of control characters for ^Z, ^[, 0xff, etc. This will make for much nicer diagnostics when working with protocols.

Documentation

data Rendering

Constructors

Rendering !Delta !Int64 !Int64 (Lines -> Lines) (Delta -> Lines -> Lines)

class HasRendering t where

Methods

rendering :: Lens t t Rendering Rendering

renderingDelta :: Lens' t Delta

renderingLine :: Lens' t (Lines -> Lines)

renderingLineBytes :: Lens' t Int64

renderingLineLen :: Lens' t Int64

renderingOverlays :: Lens' t (Delta -> Lines -> Lines)

nullRendering :: Rendering -> Bool

emptyRendering :: Rendering

class Source t where

Methods

source :: t -> (Int64, Int64, Lines -> Lines)

rendered :: Source s => Delta -> s -> Rendering

create a drawing surface

class Renderable t where

Methods

render :: t -> Rendering

data Rendered a

Constructors

a :@ Rendering

Carets

data Caret

 In file included from baz.c:9
 In file included from bar.c:4
 foo.c:8:36: note
 int main(int argc, char ** argv) { int; }
                                    ^

Constructors

Caret !Delta !ByteString

class HasCaret t where

Methods

caret :: Lens' t Caret

data Careted a

Constructors

a :^ Caret

drawCaret :: Delta -> Delta -> Lines -> Lines

addCaret :: Delta -> Rendering -> Rendering

caretEffects :: [SGR]

renderingCaret :: Delta -> ByteString -> Rendering

Spans

data Span

Constructors

Span !Delta !Delta !ByteString

class HasSpan t where

Methods

span :: Lens' t Span

data Spanned a

Constructors

a :~ Span

spanEffects :: [SGR]

drawSpan :: Delta -> Delta -> Delta -> Lines -> Lines

addSpan :: Delta -> Delta -> Rendering -> Rendering

 int main(int argc, char ** argv) { int; }
                                    ^~~

Fixits

data Fixit

Constructors

Fixit

Fields

_fixitSpan :: !Span

_fixitReplacement :: !ByteString

class HasFixit t where

Methods

fixit :: Lens t t Fixit Fixit

fixitReplacement :: Lens' t ByteString

fixitSpan :: Lens' t Span

drawFixit :: Delta -> Delta -> String -> Delta -> Lines -> Lines

addFixit :: Delta -> Delta -> String -> Rendering -> Rendering

Drawing primitives

type Lines = Array (Int, Int64) ([SGR], Char)

draw :: [SGR] -> Int -> Int64 -> String -> Lines -> Lines

ifNear :: Delta -> (Lines -> Lines) -> Delta -> Lines -> Lines

(.#) :: (Delta -> Lines -> Lines) -> Rendering -> Rendering