Text.Trifecta.Result

Description

Results and Parse Errors

Parse Results

data Result a

Constructors

Success a

Failure Doc

class AsResult p f s t a b | s -> a, t -> b, s b -> t, t a -> s where

Methods

_Result :: Overloaded p f s t (Result a) (Result b)

_Success :: (AsResult p f s t a b, Choice p, Applicative f) => Overloaded p f s t a b

_Failure :: (AsResult p f s s a a, Choice p, Applicative f) => Overloaded' p f s Doc

Parsing Errors

data Err

Constructors

Err

Fields

_reason :: Maybe Doc

_footnotes :: [Doc]

_expected :: Set String

class HasErr t where

Methods

err :: Lens' t Err

expected :: Lens' t (Set String)

footnotes :: Lens' t [Doc]

reason :: Lens' t (Maybe Doc)

explain :: Rendering -> Err -> Doc

failing :: String -> Err