Text.Trifecta.Rope

Description

Documentation

data Rope

Constructors

Rope !Delta !(FingerTree Delta Strand)

rope :: FingerTree Delta Strand -> Rope

data Strand

Constructors

Strand !ByteString !Delta

LineDirective !ByteString !Int64

strand :: ByteString -> Strand

strands :: Rope -> FingerTree Delta Strand

grabRest :: Delta -> Rope -> r -> (Delta -> ByteString -> r) -> r

grab a the contents of a rope from a given location up to a newline

grabLine :: Delta -> Rope -> r -> (Delta -> ByteString -> r) -> r

grab a the contents of a rope from a given location up to a newline