Text.Trifecta.Util.It

Description

harder, better, faster, stronger...

Documentation

data It r a

Constructors

Pure a

It a (r -> It r a)

needIt :: a -> (r -> Maybe a) -> It r a

wantIt :: a -> (r -> (#Bool, a#)) -> It r a

simplifyIt :: It r a -> r -> It r a

runIt :: (a -> o) -> (a -> (r -> It r a) -> o) -> It r a -> o

fillIt :: r -> (Delta -> ByteString -> r) -> Delta -> It Rope r

Given a position, go there, and grab the text forward from that point

rewindIt :: Delta -> It Rope (Maybe ByteString)

Return the text of the line that contains a given position

sliceIt :: Delta -> Delta -> It Rope ByteString