Text.Trifecta.Combinators

Description

Documentation

class (MonadPlus m, TokenParsing m) => DeltaParsing m where

This class provides parsers with easy access to:

1) the current line contents. 2) the current position as a Delta. 3) the ability to use sliced on any parser.

Minimal complete definition

line, position, slicedWith

Methods

line :: m ByteString

Retrieve the contents of the current line (from the beginning of the line)

position :: m Delta

Retrieve the current position as a Delta.

slicedWith :: (a -> ByteString -> r) -> m a -> m r

Run a parser, grabbing all of the text between its start and end points

rend :: m Rendering

Retrieve a Rendering of the current line noting this position, but not placing a Caret there.

restOfLine :: m ByteString

Grab the remainder of the current line

sliced :: DeltaParsing m => m a -> m ByteString

Run a parser, grabbing all of the text between its start and end points and discarding the original result

careting :: DeltaParsing m => m Caret

Grab a Caret pointing to the current location.

careted :: DeltaParsing m => m a -> m (Careted a)

Parse a Careted result. Pointing the Caret to where you start.

spanning :: DeltaParsing m => m a -> m Span

Discard the result of a parse, returning a Span from where we start to where it ended parsing.

spanned :: DeltaParsing m => m a -> m (Spanned a)

Parse a Spanned result. The Span starts here and runs to the last position parsed.

fixiting :: DeltaParsing m => m ByteString -> m Fixit

Grab a fixit.

class (DeltaParsing m, HasDelta d) => MarkParsing d m | m -> d where

This class is a refinement of DeltaParsing that adds the ability to mark your position in the input and return there for further parsing later.

Methods

mark :: m d

mark the current location so it can be used in constructing a span, or for later seeking

release :: d -> m ()

Seek a previously marked location