Text.Trifecta.Rendering

Description

The type for Lines will very likely change over time, to enable drawing lit up multi-character versions of control characters for ^Z, ^[, 0xff, etc. This will make for much nicer diagnostics when working with protocols.

Documentation

data Rendering

A Rendering is a canvas of text that output can be written to.

Constructors

Rendering !Delta !Int64 !Int64 (Lines -> Lines) (Delta -> Lines -> Lines)

class HasRendering c where

Minimal complete definition

rendering

Methods

rendering :: Lens' c Rendering

renderingDelta :: Lens' c Delta

renderingLine :: Lens' c (Lines -> Lines)

renderingLineBytes :: Lens' c Int64

renderingLineLen :: Lens' c Int64

renderingOverlays :: Lens' c (Delta -> Lines -> Lines)

nullRendering :: Rendering -> Bool

Is the Rendering empty?

>>> nullRendering emptyRendering
True

>>> nullRendering exampleRendering
False

emptyRendering :: Rendering

The empty Rendering, which contains nothing at all.

>>> show (prettyRendering emptyRendering)
""

prettyRendering :: Rendering -> Doc AnsiStyle

class Source t where

Methods

source :: t -> (Int64, Int64, Lines -> Lines)

( Number of (padded) columns
, number of bytes
, line )

rendered :: Source s => Delta -> s -> Rendering

create a drawing surface

class Renderable t where

Methods

render :: t -> Rendering

data Rendered a

Constructors

a :@ Rendering

gutterEffects :: [SGR]

ANSI terminal style for rendering the gutter.

Carets

data Caret

A Caret marks a point in the input with a simple ^ character.

>>> unAnnotate (prettyRendering (addCaret (Columns 35 35) exampleRendering))
1 | int main(int argc, char ** argv) { int; }<EOF>
  |                                    ^

Constructors

Caret !Delta !ByteString

class HasCaret t where

Methods

caret :: Lens' t Caret

data Careted a

Constructors

a :^ Caret

drawCaret :: Delta -> Delta -> Lines -> Lines

addCaret :: Delta -> Rendering -> Rendering

Render a caret at a certain position in a Rendering.

caretEffects :: [SGR]

ANSI terminal style for rendering the caret.

renderingCaret :: Delta -> ByteString -> Rendering

Spans

data Span

A Span marks a range of input characters. If Caret is a point, then Span is a line.

>>> unAnnotate (prettyRendering (addSpan (Columns 35 35) (Columns 38 38) exampleRendering))
1 | int main(int argc, char ** argv) { int; }<EOF>
  |                                    ~~~

Constructors

Span !Delta !Delta !ByteString

class HasSpan t where

Methods

span :: Lens' t Span

data Spanned a

Annotate an arbitrary piece of data with a Span, typically its corresponding input location.

Constructors

a :~ Span

spanEffects :: [SGR]

ANSI terminal style to render spans with.

drawSpan

Arguments

:: Delta
Start of the region of interest

-> Delta
End of the region of interest

-> Delta
Currrent location

-> Lines
Lines to add the rendering to

-> Lines

addSpan :: Delta -> Delta -> Rendering -> Rendering

Fixits

data Fixit

A Fixit is a Span with a suggestion.

>>> unAnnotate (prettyRendering (addFixit (Columns 35 35) (Columns 38 38) "Fix this!" exampleRendering))
1 | int main(int argc, char ** argv) { int; }<EOF>
  |                                    ~~~
  |                                    Fix this!

Constructors

Fixit

Fields

_fixitSpan :: !Span

Span where the error occurred

_fixitReplacement :: !ByteString

Replacement suggestion

class HasFixit c where

Minimal complete definition

fixit

Methods

fixit :: Lens' c Fixit

fixitReplacement :: Lens' c ByteString

fixitSpan :: Lens' c Span

drawFixit :: Delta -> Delta -> String -> Delta -> Lines -> Lines

addFixit :: Delta -> Delta -> String -> Rendering -> Rendering

Drawing primitives

type Lines = Array (Int, Int64) ([SGR], Char)

A raw canvas to paint ANSI-styled characters on.

draw

Arguments

:: [SGR]
ANSI style to use

-> Int
Line; 0 is at the top

-> Int64
Column; 0 is on the left

-> String
Data to be written

-> Lines
Canvas to draw on

-> Lines

ifNear

Arguments

:: Delta
Position 1

-> (Lines -> Lines)
Modify the fallback result if the positions are near each other

-> Delta
Position 2

-> Lines
Fallback result if the positions are not near each other

-> Lines

(.#) :: (Delta -> Lines -> Lines) -> Rendering -> Rendering