Text.Trifecta.Result

Description

Results and Parse Errors

Parse Results

data Result a

The result of parsing. Either we succeeded or something went wrong.

Constructors

Success a

Failure ErrInfo

class AsResult s t a b | s -> a, t -> b, s b -> t, t a -> s where

A Prism that lets you embed or retrieve a Result in a potentially larger type.

Methods

_Result :: Prism s t (Result a) (Result b)

foldResult :: (ErrInfo -> b) -> (a -> b) -> Result a -> b

Fold over a Result

_Success :: AsResult s t a b => Prism s t a b

The Prism for the Success constructor of Result

_Failure :: AsResult s s a a => Prism' s ErrInfo

The Prism for the Failure constructor of Result

Parsing Errors

data Err

This is used to report an error. What went wrong, some supplemental docs and a set of things expected at the current location. This does not, however, include the actual location.

Constructors

Err

Fields

_reason :: Maybe (Doc AnsiStyle)

_footnotes :: [Doc AnsiStyle]

_expected :: Set String

_finalDeltas :: [Delta]

class HasErr c where

Minimal complete definition

err

Methods

err :: Lens' c Err

expected :: Lens' c (Set String)

finalDeltas :: Lens' c [Delta]

footnotes :: Lens' c [Doc AnsiStyle]

reason :: Lens' c (Maybe (Doc AnsiStyle))

class Errable m where

Methods

raiseErr :: Err -> m a

data ErrInfo

Constructors

ErrInfo

Fields

_errDoc :: Doc AnsiStyle

_errDeltas :: [Delta]

explain :: Rendering -> Err -> Doc AnsiStyle

Convert a Rendering of auxiliary information and an Err into a 'Doc AnsiStyle', ready to be prettyprinted to the user.

failed :: String -> Err

Generate a simple Err word-wrapping the supplied message.